fix: moved url logic to helper
